Add Python 3 support
This was intially generated by the automatic `2to3` converter, and then
cleaned up by hand. It should be fully compatible with both Python 2.7
and Python 3+ (I tested on both).

print() is now a function, octal literals need an 'o' in them, and
subprocess.check_output() returns bytes, which needs to be decoded into
a str before comparison.